requirements.txt
pymessenger4
starlette==0.13.6
fastapi
uvicorn
python-decouple
pytest
requests
requests-toolbelt
six
